Revert "Merge branch 'main' into duy-2.0"
This reverts commit 96106df287399f5ba1e600f75811f9499240c2cc, reversing
changes made to cb3d30fd29ebef0d70ac569ed99f9591f1f91b9b.

+++ BAL/ViewModels/FieldTripViewModel.cs
@@ -74,8 +74,8 @@ namespace BAL.ViewModels
         public int? NumberOfParticipantsLimit { get; set; }
         [Required(ErrorMessage = "Fee is required")]
         [DisplayName("Fee")]
-        [Range(5000, int.MaxValue, ErrorMessage = "Fee must be at least 5000đ")]
-        public int? Fee { get; set; }
+        [Range(5000, double.MaxValue, ErrorMessage = "Fee must be at least 5000đ")]
+        public decimal? Fee { get; set; }
         [DisplayName("Participant Number")]
         public int? ParticipationNo { get; set; }
         [DisplayName("Review")]
